requirements.txt
langgraph==0.1.3
langchain-openai==0.1.7
supabase==2.4.0
playwright==1.44.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
sentence-transformers==3.0.1
